+++ crates/wp-oml/src/core/evaluator/query/sql.rs
@@ -147,7 +147,11 @@ impl FieldExtractor for SqlQuery {
                 out
             }
             _ => {
-                error_data!("not support more 9 params in sql eval: {}", params.len());
+                error_edata!(
+                    dst.id,
+                    "not support more 9 params in sql eval: {}",
+                    params.len()
+                );
                 //unimplemented!("not support more 9 params len ")
                 Vec::new()
             }