+++ crates/wp-oml/src/core/evaluator/transform/fmt.rs
@@ -36,8 +36,9 @@ impl FieldExtractor for FmtOperation {
                 diagnostics::push(OmlIssue::new(OmlIssueKind::FmtVarMissing, name));
             }
         }
-        debug_data!("fmt:{}, val:{:?}", self.fmt_str(), args);
-        debug_data!(
+        debug_edata!(dst.id, "fmt:{}, val:{:?}", self.fmt_str(), args);
+        debug_edata!(
+            dst.id,
             " oml fmt not get data from : {}, vars:{:?}",
             dst,
             not_find_items